Implement functionality of DisplayController
This commit implements the DisplayController functionality and
subsequently allows us to properly play the tic tac toe game.
The logic for the game itself if virtually finished (pending any bugs I
may catch in the future). All that is left to do is to clean up the
code, implement some QoL features, and play around with some of the
styling.